requirements.txt
python-telegram-bot==20.3
qrcode==7.3
requests==2.31.0